+++ packages/nav-frontend-skjema/src/checkbox-gruppe.tsx
@@ -1,0 +1,20 @@
+import * as React from "react";
+import * as classNames from "classnames";
+import { SkjemaGruppe, SkjemaGruppeProps } from ".";
+import "nav-frontend-skjema-style";
+
+class CheckboxGruppe extends React.Component<SkjemaGruppeProps> {
+  render() {
+    const { children, className, ...other } = this.props;
+    return (
+      <SkjemaGruppe
+        className={classNames("checkboxgruppe", className)}
+        {...other}
+      >
+        {children}
+      </SkjemaGruppe>
+    );
+  }
+}
+
+export default CheckboxGruppe;
